requirements.txt
bs4
fonttools
matplotlib
nltk
numpy
opencv-python-headless
pandas
pyarrow
pygame
pyyaml
scipy
seqeval
sklearn
spacy
submitit
transformers==4.17.0
wandb
